requirements.txt
telebot~=0.0.5
mysql-connector-python~=8.0.31
requests~=2.31.0
loguru=0.7.2
blinker==1.6.3
click==8.1.7
pytest==7.4.3
Faker==18.13.0
factory-boy==3.3.0
pytest-cov==4.1.0
psycopg2-binary==2.9.9
stubs==1.0.0
types-requests==2.31.0.10
typing~=3.7.4.3
loguru~=0.7.2
python-dotenv~=1.0.1